Move findPoolByLpToken to rewardsChef.helpers.ts
